Improve JS Integration
* Change JS module type
* Make seed an integer where possible
* Create package directory as part of build
* Use Sync instead of Copy for jsPackage task to prevent old files from sticking around
* Add JS runner class

+++ game-engine-core/src/main/za/co/entelect/challenge/game/engine/map/WormsMapGenerator.kt
@@ -7,9 +7,9 @@ import za.co.entelect.challenge.game.engine.powerups.Powerup
 import za.co.entelect.challenge.game.engine.simplexNoise.SimplexNoise
 import kotlin.math.*
 
-class WormsMapGenerator(private val config: GameConfig, private val seed: Long) {
+class WormsMapGenerator(private val config: GameConfig, private val seed: Int) {
 
-    private val noise = SimplexNoise(seed.toInt())
+    private val noise = SimplexNoise(seed)
     private val mapZoom = 0.3
     private val amountOfSoil = 0.50
     private var mapCenter: Pair<Double, Double> = Pair(0.0, 0.0)
